[FIX!!!] fix duration calculation so that spikeHist is correct.

+++ src/spikeSort/createBinEdge.m
@@ -3,4 +3,4 @@ function [binEdges, binEdgesPrecise] = createBinEdge(duration, samplingRate)
 % duration should be in seconds and will be converted to milliseconds.
 
 binEdges = 0:3:1000*(duration)+3;
-binEdgesPrecise = 0:2000/samplingRate:1000*(duration)+1;
+binEdgesPrecise = single(0:2000/samplingRate:1000*(duration)+1);
